mkdir to save video data

+++ pro/cgi-arm/function.cpp
@@ -14,8 +14,12 @@
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <sys/stat.h>
+#include <sys/types.h>
 #include <string.h>
 #include <fcntl.h>
+#include <mqueue.h>
+#include <dirent.h>
+
 
 extern "C"
 {
@@ -30,7 +34,9 @@ extern "C"
 int Demo_AlarmFortify();
 int saveData();
 int saveFlag = 0;
-char fileName[128] = "/mnt/sdcard/test.mp4";
+char* GetIP();
+//char fileName[128] = "/mnt/sdcard/test.mp4";
+char fileName[128] = "/test.mp4";
 LONG lUserID;
 NET_DVR_DEVICEINFO_V30 struDevice;
  long lRealPlayHandle;
@@ -84,6 +90,21 @@ void CALLBACK MessageCallback(LONG lCommand, NET_DVR_ALARMER *pAlarmer, char *pA
 
 }
 
+int mk_dir(char *dir)
+{
+    DIR *mydir = NULL;
+    if((mydir= opendir(dir))==NULL)//判断目录
+    {
+      int ret = mkdir(dir, S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);//创建目录
+      if (ret != 0)
+      {
+          return -1;
+      }
+    }
+
+    return 0;
+}
+
 void TestRebootDVR(LONG lUserID)
 {
 	if(NET_DVR_RebootDVR(lUserID))
@@ -104,23 +125,9 @@ void TestShutdownDVR(LONG lUserID)
 int saveData()
 {
 	/*
-	 long lRealPlayHandle;
-    NET_DVR_CLIENTINFO ClientInfo = {0};
-#if (defined(_WIN32) || defined(_WIN_WCE)) || defined(__APPLE__)
-    ClientInfo.hPlayWnd     = NULL;
-#elif defined(__linux__)
-   ClientInfo.hPlayWnd     = 0;
-#endif
-
-    ClientInfo.lChannel     = 1;  //channel NO.
-    //ClientInfo.lLinkMode    = 0x40000000; //Record when breaking network.
-    ClientInfo.lLinkMode    = 0;
-    ClientInfo.sMultiCastIP = NULL;
-*/
 	printf("Start realplay...\n");
-    //getchar();
-    //lRealPlayHandle = NET_DVR_RealPlay_V30(lUserID, &ClientInfo, g_RealDataCallBack_V30, NULL, 0);
-    //lRealPlayHandle = NET_DVR_RealPlay_V30(lUserID, &ClientInfo,NULL,NULL,0);
+    lRealPlayHandle = NET_DVR_RealPlay_V30(lUserID, &ClientInfo, g_RealDataCallBack_V30, NULL, 0);
+    lRealPlayHandle = NET_DVR_RealPlay_V30(lUserID, &ClientInfo,NULL,NULL,0);
     if (lRealPlayHandle < 0)
     {
         printf("pyd1---NET_DVR_RealPlay_V30 error\n");
@@ -129,20 +136,25 @@ int saveData()
 
     //Set rtsp callback function of getting stream.
     //NET_DVR_SetStandardDataCallBack(lRealPlayHandle, g_StdDataCallBack, 0);
+    */
     
+    char dir[128]="/opt/homewell/video/";
+    strcat(dir,GetIP());
+    mk_dir(dir);
     printf("Start save data...\n");
     time_t nowtime;
 	struct tm *timeinfo;
 
 	time(&nowtime);
 	timeinfo = localtime(&nowtime);
-	sprintf(fileName,"%d%02d%02d-%02d-%02d-%02d.mp4",1900+timeinfo->tm_year,1+timeinfo->tm_mon,timeinfo->tm_mday,timeinfo->tm_hour,timeinfo->tm_min,timeinfo->tm_sec);
-	char filePath[64] = "/mnt/sdcard/";
-	strcat(filePath,fileName);
-    int iRet;
-	//fd = open(filePath,O_RDWR|O_CREAT|O_TRUNC,S_IRWXU|S_IRGRP|S_IROTH);
-    //close(fd);
-	iRet = NET_DVR_SaveRealData(lRealPlayHandle,filePath);
+	sprintf(fileName,"%s%d%02d%02d-%02d-%02d-%02d.mp4","/",1900+timeinfo->tm_year,1+timeinfo->tm_mon,timeinfo->tm_mday,timeinfo->tm_hour,timeinfo->tm_min,timeinfo->tm_sec);
+	//char filePath[64] = "/mnt/sdcard/";
+	//char filePath[64] = "./";
+	strcat(dir,fileName);
+    int iRet,fd;
+	fd = open(dir,O_RDWR|O_CREAT|O_TRUNC,S_IRWXU|S_IRGRP|S_IROTH);
+    close(fd);
+	iRet = NET_DVR_SaveRealData(lRealPlayHandle,dir);
     if(!iRet)
     {
 		printf("pyd1---NET_DVR_SaveRealData error:%s\n",NET_DVR_GetLastError());
@@ -231,10 +243,17 @@ int capture()
     strPicPara.wPicQuality = 2;
     strPicPara.wPicSize = 0;
     int iRet,fd;
-	char file[64] = "/mnt/sdcard/temp.jpeg";
-    fd = open(file,O_RDWR|O_TRUNC|O_CREAT|O_APPEND,S_IRWXU|S_IRGRP|S_IROTH);
+    
+    char dir[128]="/opt/homewell/video/";
+    strcat(dir,GetIP());
+    mk_dir(dir);
+	//char file[64] = "/mnt/sdcard/temp.jpeg";
+	char file[64] = "/temp.jpeg";
+	strcat(dir,file);
+    fd = open(dir,O_RDWR|O_TRUNC|O_CREAT|O_APPEND,S_IRWXU|S_IRGRP|S_IROTH);
 	close(fd);
-	iRet = NET_DVR_CaptureJPEGPicture(lUserID, struDevice.byStartChan, &strPicPara, file);
+	printf("dir is %s\n",dir);
+	iRet = NET_DVR_CaptureJPEGPicture(lUserID, struDevice.byStartChan, &strPicPara, dir);
     if (!iRet)
     {
         printf("pyd1---NET_DVR_CaptureJPEGPicture error, %d\n", NET_DVR_GetLastError());
@@ -267,11 +286,15 @@ int record()
 		return 1;
 	}
   
+	char dir[128]="/opt/homewell/video/";
+    strcat(dir,GetIP());
+    mk_dir(dir);
+    strcat(dir,fileName);
 	printf("Start save...\n");
 	int iRet,fd;	
-	fd = open(fileName,O_RDWR|O_CREAT|O_TRUNC|O_APPEND,S_IRWXU|S_IRGRP|S_IROTH);//先调用open创建之后，不许要root权限即可录制，避免文加大小为0
+	fd = open(dir,O_RDWR|O_CREAT|O_TRUNC|O_APPEND,S_IRWXU|S_IRGRP|S_IROTH);//先调用open创建之后，不许要root权限即可录制，避免文加大小为0
 	close(fd);
-	iRet = NET_DVR_SaveRealData(lRealPlayHandle,fileName);
+	iRet = NET_DVR_SaveRealData(lRealPlayHandle,dir);
     if(!iRet)
     {
 		printf("pyd1---NET_DVR_SaveRealData error:%s\n",NET_DVR_GetLastError());
@@ -283,12 +306,101 @@ int record()
 	printf("saving...\n");
 	sleep(10);
 	
+	NET_DVR_StopSaveRealData(lRealPlayHandle);
 	printf("save data success!\n");
 	return HPR_OK;
 }
+
+char *GetIP()
+{
+	/*
+    mqd_t mqd;  
+    struct mq_attr attr;  
+    char *ptr;  
+    unsigned int prio;  
+    size_t n;  
+    int rc;  
+	const char *file = "/getCameraIP";
+     
+
+    //只读模式打开消息队列
+    mqd = mq_open(file, O_RDONLY);  
+    if(mqd < 0)  
+    {  
+        perror("打开消息队列失败");  
+        return NULL;   
+    }     
+
+    // 取得消息队列属性，根据mq_msgsize动态申请内存  
+    rc = mq_getattr(mqd, &attr);  
+    if(rc < 0)  
+    {  
+        perror("取得消息队列属性失败");  
+        return NULL;   
+    }  
+
+    //动态申请保证能存放单条消息的内存 
+    ptr = (char*)calloc(attr.mq_msgsize, sizeof(char));  
+    if(NULL == ptr)  
+    {  
+        printf("动态申请内存失败\n");  
+        mq_close(mqd);  
+        return NULL;   
+    }     
+	printf("size %ld\n",attr.mq_msgsize);
+    //*接收一条消息
+    n = mq_receive(mqd, ptr, attr.mq_msgsize, &prio);  
+    if(n < 0)  
+    {  
+        perror("读取失败");  
+        mq_close(mqd);  
+        free(ptr);  
+        return NULL;   
+    }  
+    
+    printf("received %s \n", ptr); 
+    rc = mq_close(mqd);  
+    if(0 != rc)  
+    {  
+        perror("关闭失败");  
+        return NULL;  
+    }    
+    
+    rc = mq_unlink(file);  
+    if(0 != rc)  
+    {  
+        perror("关闭失败");  
+        return NULL;  
+    } 
+    */
+    char *ip,*file = "/tmp/cameraIP";
+    int fd,len = 32;
+    
+    ip = (char*)calloc(len,sizeof(char));
+    if(fd == -1)
+    {
+		perror("calloc");
+		return NULL;
+	}
+	
+    fd = open(file,O_RDONLY,0644);
+    if(fd == -1)
+    {
+		perror("open");
+		return NULL;
+	}
+	
+	read(fd,ip,len);
+	
+	close(fd);
+    return ip;  
+}
+
 int cgiMain(void)
 {
-
+	
+	char *ip;
+	ip = (char *) calloc(32, sizeof(char));
 	//初始化
 	NET_DVR_Init();
 
@@ -301,7 +413,8 @@ int cgiMain(void)
 	//注册设备
 	LONG lUserID;
 	//NET_DVR_DEVICEINFO_V30 struDevice;
-	lUserID = NET_DVR_Login_V30("192.168.1.64",8000,"admin","12345",&struDevice);
+	strcpy(ip,GetIP());
+	lUserID = NET_DVR_Login_V30(ip,8000,"admin","12345",&struDevice);
 	if(lUserID < 0)
 	{
 		printf("Login in ERROR:%d\n",NET_DVR_GetLastError());
@@ -309,7 +422,12 @@ int cgiMain(void)
 		return 1;
 	}
 	printf("login success!\n");
-
+	free(ip);
+	
+	//---------------------------------------
+	//设置异常消息回调函数
+	NET_DVR_SetExceptionCallBack_V30(0, NULL,g_ExceptionCallBack, NULL);
+	
 	NET_DVR_CLIENTINFO ClientInfo = {0};
 #if (defined(_WIN32) || defined(_WIN_WCE)) || defined(__APPLE__)
     ClientInfo.hPlayWnd     = NULL;
@@ -321,7 +439,17 @@ int cgiMain(void)
     //ClientInfo.lLinkMode    = 0x40000000; //Record when breaking network.
     ClientInfo.lLinkMode    = 0;
     ClientInfo.sMultiCastIP = NULL;
+    BOOL bPreviewBlock = false;       //请求码流过程是否阻塞，0：否，1：是
 	lRealPlayHandle = NET_DVR_RealPlay_V30(lUserID, &ClientInfo,NULL,NULL,0);
+	if (lRealPlayHandle < 0)
+	{
+		printf("NET_DVR_RealPlay_V30 error:%d\n",NET_DVR_GetLastError());
+		NET_DVR_Logout(lUserID);
+		NET_DVR_Cleanup();
+		return 1;
+	}
+	
+	
 	cgiHeaderContentType("Text/html");
 	//while(1){
 		if(cgiFormSubmitClicked("reboot") == cgiFormSuccess)
